refactor: move appliesToRoute into new TemporaryTerminalRewriter
